Fix type of return

+++ 0x00-python_variable_annotations/2-floor.py
@@ -3,7 +3,7 @@
 import math
 
 
-def floor(n: float) -> float:
+def floor(n: float) -> int:
     """[Function that takes a float n as argument and
         returns the floor of the float]
 
@@ -11,6 +11,6 @@ def floor(n: float) -> float:
             n (float): [number used in operation]
 
         Returns:
-            float: [it's the result of calculate the floor of number]
+            int: [it's the result of calculate the floor of number]
         """
     return math.floor(n)
